Disaster recovery system for virtual machines using VMware API for ESXi 5.0 & vCenter server  Disaster recovery system for virtual machines
The Availability Manger provides a solution that manages the Virtual
Machines in a virtual infrastructure and also recovers them in case of
any failure event. It is also capable of adding new Virtual Hosts and
Virtual Machines when these machines fail and are not able to reconnect
to the network.